Adding weights to Graph

+++ leafy/data_structure.pxd
@@ -24,7 +24,7 @@ cdef class AdjacencyList:
 
 
 cdef class MemoryViewArrayIter:
-    cdef int [::1] _mv_array
+    cdef double [::1] _mv_array
     cdef int _length
     cdef int _counter
